refactor(FE): `useRegister` hook
