+++ frontend/src/hooks/mutations.js
@@ -4,25 +4,6 @@ import { QueryKeys } from "../utils/utils";
 import useBlockchain from "./useBlockchain";
 import useToast from "./useToast";
 
-export function useRegister() {
-  const { data: blockchain } = useBlockchain();
-  const queryClient = useQueryClient();
-
-  return useMutation(
-    async () => {
-      const tx = await blockchain.fakemon.registerUser();
-      await tx.wait();
-      queryClient.setQueryData([QueryKeys.IS_REGISTERED], true);
-    },
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries([QueryKeys.TOKEN_BALANCE]);
-        queryClient.invalidateQueries([QueryKeys.FAKEMONS]);
-      },
-    }
-  );
-}
-
 export function useMintFakemon() {
   const { data: blockchain } = useBlockchain();
   const queryClient = useQueryClient();
