+++ frontend/src/hooks/useRegister.js
@@ -1,0 +1,24 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { QueryKeys } from "../utils/utils";
+import useBlockchain from "./useBlockchain";
+
+function useRegister() {
+  const { data: blockchain } = useBlockchain();
+  const queryClient = useQueryClient();
+
+  return useMutation(
+    async () => {
+      const tx = await blockchain.fakemon.registerUser();
+      await tx.wait();
+      queryClient.setQueryData([QueryKeys.IS_REGISTERED], true);
+    },
+    {
+      onSuccess: () => {
+        queryClient.invalidateQueries([QueryKeys.TOKEN_BALANCE]);
+        queryClient.invalidateQueries([QueryKeys.FAKEMONS]);
+      },
+    }
+  );
+}
+
+export default useRegister;
